requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-dotenv==1.0.1
openai==1.30.1
httpx==0.27.0
alembic==1.13.1
tabulate==0.9.0
rich==13.7.1
